# public/swagger.yaml
openapi: 3.0.3
info:
  title: Vinci Barbearia API
  description: |
    API completa para gerenciamento de barbearia incluindo agendamentos, barbeiros, clientes, produtos, planos e webhooks.

    ## Autenticação
    A maioria dos endpoints requer autenticação via Bearer Token no header:
    ```
    Authorization: Bearer SEU_TOKEN_AQUI
    ```

    ## Formato de Datas
    - Datas são no formato brasileiro: DD/MM/YYYY ou YYYY-MM-DD
    - Horários são no formato: HH:MM (24h)

    ## Webhooks
    O sistema suporta dois tipos de webhooks:
    1. **Webhook Global**: Configurado na tabela `configuracoes`, dispara para todos os eventos
    2. **Webhooks de Barbeiros**: Webhooks personalizados por profissional

  version: 3.0.0
  contact:
    name: Vinci Barbearia
    url: https://vincebarbearia.vercel.app
servers:
  - url: https://vincebarbearia.vercel.app/api
    description: Produção
  - url: http://localhost:3000/api
    description: Desenvolvimento
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token de autenticação da API
...
paths:
...
  /barbeiros/faturamento-mes:
    get:
      tags:
        - Barbeiros
      summary: Faturamento do barbeiro no mês
      description: |
        Retorna o faturamento do MÊS ATUAL (ou mês especificado) para um barbeiro específico.

        **⚠️ ESCOLHA UMA das opções abaixo para identificar o barbeiro:**
        - Por telefone (campo: telefone)
        - Por nome do barbeiro (campo: barbeiro_nome)
        - Por ID do barbeiro (campo: barbeiro_id)
      parameters:
        - name: telefone
          in: query
          required: false
          schema:
            type: string
          description: "[OPÇÃO 1] Telefone do barbeiro"
          example: "5511999999999"
        - name: barbeiro_nome
          in: query
          required: false
          schema:
            type: string
          description: "[OPÇÃO 2] Nome do barbeiro"
          example: "Hiago"
        - name: barbeiro_id
          in: query
          required: false
          schema:
            type: string
            format: uuid
          description: "[OPÇÃO 3] UUID do barbeiro"
          example: "550e8400-e29b-41d4-a716-446655440000"
        - name: mes
          in: query
          schema:
            type: string
          description: "Mês no formato MM (01-12). Padrão: mês atual"
          example: "12"
        - name: ano
          in: query
          schema:
            type: string
          description: "Ano no formato YYYY. Padrão: ano atual"
          example: "2025"
      responses:
        '200':
          description: Faturamento retornado com sucesso
...
  /barbeiros/horarios:
    get:
      tags:
        - Barbeiros
      summary: Consultar horários do barbeiro
      description: Retorna configuração de horários de trabalho
      parameters:
        - name: barbeiro_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Horários retornados
...